+++ mobile/src/data/playlist/queries.ts
@@ -1,7 +1,18 @@
-import { useQuery } from "@tanstack/react-query";
+import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
 import { useTranslation } from "react-i18next";
 
+import type { playlists } from "~/db/schema";
+
+import { sanitizePlaylistName } from "~/api/playlist.utils";
+import { Resynchronize } from "~/stores/Playback/actions";
 import { queries as q } from "~/queries/keyStore";
+import { createPlaylist, deletePlaylist, updatePlaylist } from "./api";
+
+import { wait } from "~/utils/promise";
+
+type InsertedPlaylist = Omit<typeof playlists.$inferInsert, "isFavorite"> & {
+  tracks?: Array<{ id: string }>;
+};
 
 //#region Queries
 export function usePlaylist(playlistName: string) {
@@ -40,6 +51,74 @@ export function usePlaylistsNames() {
   return useQuery({
     ...q.playlists.all,
     select: (data) => data.map((p) => p.name),
+  });
+}
+//#endregion
+
+//#region Mutations
+export function useCreatePlaylist() {
+  const queryClient = useQueryClient();
+  return useMutation({
+    mutationFn: (args: Pick<InsertedPlaylist, "name" | "tracks">) =>
+      createPlaylist(args),
+    onSuccess: (_, { tracks }) => {
+      queryClient.invalidateQueries({ queryKey: q.playlists._def });
+      // Invalidate all track queries if we've added tracks.
+      if (tracks) queryClient.invalidateQueries({ queryKey: q.tracks._def });
+      queryClient.invalidateQueries({ queryKey: ["search"] });
+    },
+  });
+}
+
+export function useDeletePlaylist(playlistName: string) {
+  const queryClient = useQueryClient();
+  return useMutation({
+    mutationFn: () => deletePlaylist(playlistName),
+    onSuccess: () => {
+      queryClient.invalidateQueries({ queryKey: q.playlists._def });
+      queryClient.invalidateQueries({ queryKey: q.favorites.lists.queryKey });
+      queryClient.invalidateQueries({ queryKey: ["search"] });
+    },
+  });
+}
+
+export function useFavoritePlaylist(playlistName: string) {
+  const queryClient = useQueryClient();
+  return useMutation({
+    mutationFn: async (isFavorite: boolean) => {
+      await wait(1);
+      return updatePlaylist(playlistName, { isFavorite });
+    },
+    onSuccess: () => {
+      queryClient.invalidateQueries({
+        queryKey: q.playlists.detail(playlistName).queryKey,
+      });
+      queryClient.invalidateQueries({ queryKey: q.favorites.lists.queryKey });
+    },
+  });
+}
+
+export function useUpdatePlaylist(playlistName: string) {
+  const queryClient = useQueryClient();
+  return useMutation({
+    mutationFn: (args: Partial<InsertedPlaylist>) =>
+      updatePlaylist(playlistName, args),
+    onSuccess: async (_, { name }) => {
+      queryClient.resetQueries({ queryKey: q.playlists._def });
+      queryClient.invalidateQueries({ queryKey: q.tracks._def });
+      queryClient.invalidateQueries({ queryKey: q.favorites.lists.queryKey });
+      queryClient.invalidateQueries({ queryKey: ["search"] });
+
+      // `sanitizePlaylistName` shouldn't throw an error.
+      const sanitizedName = name ? sanitizePlaylistName(name) : undefined;
+
+      if (sanitizedName) {
+        await Resynchronize.onRename({
+          oldSource: { type: "playlist", id: playlistName },
+          newSource: { type: "playlist", id: sanitizedName },
+        });
+      }
+    },
   });
 }
 //#endregion
