+++ mobile/src/modules/backup/JSON.ts
@@ -11,11 +11,12 @@ import type { TrackWithRelations } from "~/db/schema";
 import { albums, playlists } from "~/db/schema";
 
 import i18next from "~/modules/i18n";
-import { createPlaylist, getPlaylists, updatePlaylist } from "~/api/playlist";
+import { getPlaylists } from "~/api/playlist";
 import { sanitizePlaylistName } from "~/api/playlist.utils";
 import { getTracks } from "~/api/track";
 import { TrackList } from "~/api/track.utils";
 import { getAlbums } from "~/data/album/api";
+import { createPlaylist, updatePlaylist } from "~/data/playlist/api";
 
 import { pickDirectory } from "~/lib/file-system";
 import { clearAllQueries } from "~/lib/react-query";
