--- mobile/src/queries/playlist.ts
@@ -1,104 +1,0 @@
-import { useMutation, useQueryClient } from "@tanstack/react-query";
-
-import type { playlists } from "~/db/schema";
-
-import {
-  createPlaylist,
-  deletePlaylist,
-  favoritePlaylist,
-  updatePlaylist,
-} from "~/api/playlist";
-import { sanitizePlaylistName } from "~/api/playlist.utils";
-import { Resynchronize } from "~/stores/Playback/actions";
-import { queries as q } from "./keyStore";
-
-import { wait } from "~/utils/promise";
-
-//#region Mutations
-/** Create a new playlist. */
-export function useCreatePlaylist() {
-  const queryClient = useQueryClient();
-  return useMutation({
-    mutationFn: (args: {
-      playlistName: string;
-      tracks?: Array<{ id: string }>;
-    }) => createPlaylist({ name: args.playlistName, tracks: args.tracks }),
-    onSuccess: (_, { tracks }) => {
-      queryClient.invalidateQueries({ queryKey: q.playlists._def });
-      // The `q.tracks.detail()._ctx.playlists` key needs to be updated for
-      // each track added to this playlist on creation.
-      if (tracks) {
-        tracks.forEach(({ id }) =>
-          queryClient.invalidateQueries({
-            queryKey: q.tracks.detail(id)._ctx.playlists.queryKey,
-          }),
-        );
-      }
-      queryClient.invalidateQueries({ queryKey: ["search"] });
-    },
-  });
-}
-
-/** Delete specified playlist. */
-export function useDeletePlaylist(playlistName: string) {
-  const queryClient = useQueryClient();
-  return useMutation({
-    mutationFn: () => deletePlaylist(playlistName),
-    onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: q.playlists._def });
-      queryClient.invalidateQueries({ queryKey: q.favorites.lists.queryKey });
-      queryClient.invalidateQueries({ queryKey: ["search"] });
-    },
-  });
-}
-
-/** Set the favorite status of an playlist. */
-export function useFavoritePlaylist(playlistName: string) {
-  const queryClient = useQueryClient();
-  return useMutation({
-    /** Pass the new favorite status of the playlist. */
-    mutationFn: async (isFavorite: boolean) => {
-      await wait(1);
-      await favoritePlaylist(playlistName, isFavorite);
-    },
-    onSuccess: () => {
-      queryClient.invalidateQueries({
-        queryKey: q.playlists.detail(playlistName).queryKey,
-      });
-      queryClient.invalidateQueries({ queryKey: q.favorites.lists.queryKey });
-    },
-  });
-}
-
-/** Update specified playlist. */
-export function useUpdatePlaylist(playlistName: string) {
-  const queryClient = useQueryClient();
-  return useMutation({
-    mutationFn: (
-      updatedValues: Partial<
-        Omit<typeof playlists.$inferInsert, "isFavorite">
-      > & { tracks?: Array<{ id: string }> },
-    ) => updatePlaylist(playlistName, updatedValues),
-    onSuccess: async (_, { name }) => {
-      queryClient.resetQueries({ queryKey: q.playlists._def });
-      // Need to update all track queries as we don't exactly know which
-      // were removed.
-      queryClient.invalidateQueries({ queryKey: q.tracks._def });
-      // Invalidate favorite lists query to update the artwork or name used.
-      queryClient.invalidateQueries({ queryKey: q.favorites.lists.queryKey });
-      queryClient.invalidateQueries({ queryKey: ["search"] });
-
-      // Don't need a try-catch as `name` should be valid (as it's called
-      // in `updatePlaylist()`).
-      const sanitizedName = name ? sanitizePlaylistName(name) : undefined;
-
-      if (sanitizedName) {
-        await Resynchronize.onRename({
-          oldSource: { type: "playlist", id: playlistName },
-          newSource: { type: "playlist", id: sanitizedName },
-        });
-      }
-    },
-  });
-}
-//#endregion
